+++ src/library/DIPS.Mobile.UI/Components/TextFields/Editor/Android/EditorHandler.cs
@@ -1,6 +1,5 @@
 using Android.Graphics.Drawables;
 using AndroidX.AppCompat.Widget;
-using AndroidX.Fragment.App;
 using DIPS.Mobile.UI.Components.BottomSheets.Android;
 using Microsoft.Maui.Controls.Platform;
 using Microsoft.Maui.Platform;