+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [20.8.1]
-- Fixed an issue where buttons could not take up full width
+- Fixed an issue where buttons in BottombarButtons (BottomSheet) could not take up full width
 
 ## [20.8.0]
 - Fixed an issue where changing the 'Position' property on BottomSheet would only run one time.